# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Fitnessstracker server.",
        "title": "Fitnesstracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/change-password": {
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "set new password with CurrentPassword, NewPassword",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "set new password for user",
                "operationId": "user-password-update",
                "parameters": [
                    {
                        "description": "The input for set new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserPassword"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User password successfully updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data / current password and new password can't be same",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "401": {
                        "description": "Email id doesn't exist",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "login user with Email, Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "login for a user",
                "operationId": "user-login",
                "parameters": [
                    {
                        "description": "The input for login for user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserID"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "401": {
                        "description": "Email id doesn't exist / Wrong password",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/otp/request": {
            "post": {
                "description": "send otp in registered email for set new user password in case of forgot password with Email, EventType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "otp for forgot user password",
                "operationId": "user-otprequest",
                "parameters": [
                    {
                        "description": "The input for otp for forgot password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RequestOTP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP sent to email Successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "401": {
                        "description": "Email id doesn't exist",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/otp/verify": {
            "post": {
                "description": "otp verification for otp sent in registered email for set new user password in case of forgot password with Email, EventType, OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "verify otp for forgot user password",
                "operationId": "user-verifyotp",
                "parameters": [
                    {
                        "description": "The input for verify otp for forgot password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VerifyOTP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Token"
                        }
                    },
                    "401": {
                        "description": "Invalid OTP",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "410": {
                        "description": "OTP Expired",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "add new user details with Email, FullName, Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Add a new user",
                "operationId": "user-signup",
                "parameters": [
                    {
                        "description": "The input for add new user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserSignup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserID"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/user/alldetails": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "fetch all details by date",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserDetails"
                ],
                "summary": "fetch all details",
                "operationId": "user-alldetails-fetch",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"2024-01-01\"",
                        "description": "date for which want to see details",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AllDetails"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/user/exercise/": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "update exercise details like Duration, ExerciseType, CaloriesBurned",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercise"
                ],
                "summary": "update exercise details of today",
                "operationId": "user-exercise-update",
                "parameters": [
                    {
                        "description": "The input for update the exercise details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Exercise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "add new exercise details with ExerciseType, Duration, CaloriesBurned, Date(default)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercise"
                ],
                "summary": "Add a new exercise for today",
                "operationId": "user-exercise-add",
                "parameters": [
                    {
                        "description": "The input for add exercise",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Exercise"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully added",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "delete exercise details of today with ExerciseType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exercise"
                ],
                "summary": "delete exercise details of today",
                "operationId": "user-exercise-delete",
                "parameters": [
                    {
                        "description": "The input for delete exercise",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ExerciseType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/user/meal/": {
            "put": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "update meal details like Ingredients, MealType, CaloriesConsumed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meal"
                ],
                "summary": "update meal details of today",
                "operationId": "user-meal-update",
                "parameters": [
                    {
                        "description": "The input for update the meal details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Meal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully updated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "add new meal details with MealType, Ingredients, CaloriesConsumed, Date(default)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meal"
                ],
                "summary": "Add a new meal details for today",
                "operationId": "user-meal-add",
                "parameters": [
                    {
                        "description": "The input for add meal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Meal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully added",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "delete meal details of today with MealType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meal"
                ],
                "summary": "delete meal details of today",
                "operationId": "user-meal-delete",
                "parameters": [
                    {
                        "description": "The input for delete meal",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MealType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details Successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "security": [
                    {
                        "UserIDAuth": []
                    }
                ],
                "description": "fetch user profile with Email, FullName, Age, Gender, Height, Weight, HealthGoal, ProfilePhoto",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserDetails"
                ],
                "summary": "fetch user profile details",
                "operationId": "user-profiledetails",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AllDetails"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Message": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.RequestOTP": {
            "type": "object",
            "required": [
                "email",
                "eventType"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string",
                    "enum": [
                        "forgot_password"
                    ]
                }
            }
        },
        "models.Token": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "models.UpdateUserPassword": {
            "type": "object",
            "required": [
                "currentPassword",
                "newPassword"
            ],
            "properties": {
                "currentPassword": {
                    "type": "string"
                },
                "newPassword": {
                    "type": "string"
                }
            }
        },
        "models.UserID": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.UserLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "models.VerifyOTP": {
            "type": "object",
            "required": [
                "email",
                "eventType",
                "otp"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "UserIDAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}